+++ src/app/components/input-group/page.tsx
@@ -1,0 +1,6 @@
+'use client';
+import Documentation from './input-group.mdx';
+
+export default function Page() {
+  return <Documentation />;
+}